Restructure directory layout for having multiple tests
